pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "jinpress"
version = "0.9.0"
description = "A fast, lightweight, and elegantly configured Python static site generator inspired by VitePress"
readme = "README.md"
license = "MIT"
authors = [
    {name = "JinPress Team"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Documentation",
    "Topic :: Software Development :: Documentation",
    "Topic :: Text Processing :: Markup :: HTML",
]
requires-python = ">=3.8"
dependencies = [
    "click>=8.0.0",
    "minijinja>=1.0.0",
    "markdown-it-py>=3.0.0",
    "pyyaml>=6.0.0",
    "watchdog>=3.0.0",
    "livereload>=2.6.0",
    "pygments>=2.15.0",
    "mdit-py-plugins>=0.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "hypothesis>=6.0.0",
]

[project.scripts]
jinpress = "jinpress.cli:main"

[project.urls]
Homepage = "https://github.com/slime21023/jinpress"
Documentation = "https://github.com/slime21023/jinpress/tree/main/document"
Repository = "https://github.com/slime21023/jinpress"
Issues = "https://github.com/slime21023/jinpress/issues"

[tool.hatch.build.targets.wheel]
packages = ["jinpress"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long - handled by formatter for code, but not for string templates
]

[tool.ruff.lint.per-file-ignores]
"jinpress/scaffold.py" = ["W291", "W293", "E501"]  # Allow whitespace in string templates

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[dependency-groups]
dev = [
    "build>=1.2.2.post1",
]

[tool.ruff.lint.isort]
known-first-party = ["jinpress"]
